requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
python-jose[cryptography]
python-multipart
